Add multiple lobby support
Each lobby represents an enclosed area consisting of a game and the
connected players. The lobby itself serves as a mediator between them,
following the mediator design pattern. This design provides a solid
foundation to support multiple lobbies and facilitate communication
between the game logic and the connected clients.

If no futher issues occur, this commit will resolve #16.